+++ experimental/examples/imageviewer/shared/build.gradle.kts
@@ -72,7 +72,7 @@ android {
     compileSdk = 33
     sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
     sourceSets["main"].res.srcDirs("src/androidMain/res")
-    sourceSets["main"].resources.srcDir("src/commonMain/resources")//todo check another samples
+    sourceSets["main"].resources.srcDir("src/commonMain/resources")
     defaultConfig {
         minSdk = 24
         targetSdk = 33
