+++ experimental/examples/imageviewer/shared/src/androidMain/kotlin/example/imageviewer/view/ScalableImage.android.kt
@@ -2,6 +2,7 @@ package example.imageviewer.view
 
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
+import example.imageviewer.model.ScalableState
 
 actual fun Modifier.addUserInput(state: MutableState<ScalableState>) =
     addTouchUserInput(state)
