--- experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/core/EventLocker.kt
@@ -1,18 +1,0 @@
-package example.imageviewer.core
-
-class EventLocker {
-
-    private var value: Boolean = false
-
-    fun lock() {
-        value = false
-    }
-
-    fun unlock() {
-        value = true
-    }
-
-    fun isLocked(): Boolean {
-        return value
-    }
-}
