+++ experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/model/Config.kt
@@ -1,0 +1,10 @@
+package example.imageviewer.model
+
+//todo dima-avdeev/add-uikit-to-imageviewer to master
+val BASE_URL =
+    "https://raw.githubusercontent.com/JetBrains/compose-jb/dima-avdeev/add-uikit-to-imageviewer/artwork/imageviewerrepo"
+val PICTURES_DATA_URL = "$BASE_URL/pictures.json"
+
+const val MAX_SCALE = 5f
+const val MIN_SCALE = 1f
+const val TOAST_DURATION = 3000L
