+++ experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/model/State.kt
@@ -5,40 +5,28 @@ import androidx.compose.ui.graphics.ImageBitmap
 import example.imageviewer.Dependencies
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
-import kotlinx.coroutines.*
-import kotlinx.serialization.Serializable
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.launch
 import kotlinx.serialization.builtins.ListSerializer
-import kotlinx.serialization.json.Json
-
-@Serializable
-data class Picture(val big: String, val small: String)
-
-val Picture.bigUrl get() = "$BASE_URL/$big"
-val Picture.smallUrl get() = "$BASE_URL/$small"
-
-//todo dima-avdeev/add-uikit-to-imageviewer to master
-val BASE_URL =
-    "https://raw.githubusercontent.com/JetBrains/compose-jb/dima-avdeev/add-uikit-to-imageviewer/artwork/imageviewerrepo"
-
-val PICTURES_DATA_URL = "$BASE_URL/pictures.json"
-val jsonReader = Json {
-    ignoreUnknownKeys = true
-}
-
-sealed interface ScreenState {
-    object Miniatures : ScreenState
-    object FullScreen : ScreenState
-}
 
 data class State(
-    val mainImage: ImageBitmap? = null,
     val currentImageIndex: Int = 0,
     val miniatures: Map<Picture, ImageBitmap> = emptyMap(),
     val pictures: List<Picture> = emptyList(),
     val screen: ScreenState = ScreenState.Miniatures
 )
 
-fun MutableState<State>.modifyState(modification: State.() -> State) {
+sealed interface ScreenState {
+    object Miniatures : ScreenState
+    object FullScreen : ScreenState
+}
+
+val State.isContentReady get() = pictures.isNotEmpty()
+val State.picture get():Picture? = pictures.getOrNull(currentImageIndex)
+
+fun <T> MutableState<T>.modifyState(modification: T.() -> T) {
     value = value.modification()
 }
 
@@ -61,7 +49,7 @@ fun MutableState<State>.previousImage() = modifyState {
 fun MutableState<State>.refresh(dependencies: Dependencies) {
     dependencies.ioScope.launch {
         try {
-            val pictures = jsonReader.decodeFromString(
+            val pictures = dependencies.json.decodeFromString(
                 ListSerializer(Picture.serializer()),
                 dependencies.httpClient.get(PICTURES_DATA_URL).bodyAsText()
             )
@@ -85,72 +73,13 @@ fun MutableState<State>.refresh(dependencies: Dependencies) {
     }
 }
 
-fun MutableState<State>.setMainImage(picture: Picture, dependencies: Dependencies) {
-    dependencies.ioScope.launch {
-        val mainImage = dependencies.imageRepository.loadContent(picture.bigUrl)
-        modifyState {
-            copy(mainImage = mainImage)
-        }
-    }
+fun MutableState<State>.setSelectedIndex(index: Int) = modifyState {
+    copy(currentImageIndex = index)
 }
 
-val State.isContentReady get() = pictures.isNotEmpty()
-val State.picture get():Picture? = pictures.getOrNull(currentImageIndex)
-
-interface Notification {
-    fun notifyInvalidRepo()
-    fun notifyRepoIsEmpty()
-    fun notifyNoInternet()
-    fun notifyLoadImageUnavailable()
-    fun notifyLastImage()
-    fun notifyFirstImage()
-    fun notifyImageData(picture: Picture)
-    fun notifyRefreshUnavailable()
-}
-
-abstract class PopupNotification(val localization: Localization) : Notification {
-    abstract fun showPopUpMessage(text: String)
-
-    override fun notifyInvalidRepo() = showPopUpMessage(localization.repoInvalid)
-    override fun notifyRepoIsEmpty() = showPopUpMessage(localization.repoEmpty)
-    override fun notifyNoInternet() = showPopUpMessage(localization.noInternet)
-    override fun notifyLoadImageUnavailable() =
-        showPopUpMessage(
-            """
-                ${localization.noInternet}
-                ${localization.loadImageUnavailable}
-            """.trimIndent()
-        )
-
-    override fun notifyLastImage() = showPopUpMessage(localization.lastImage)
-    override fun notifyFirstImage() = showPopUpMessage(localization.firstImage)
-    override fun notifyImageData(picture: Picture) = showPopUpMessage(
-        """
-            ${localization.picture} ${picture.name}
-            ${localization.size} ${picture.width}x${picture.height} ${localization.pixels}
-        """.trimIndent()
+fun MutableState<State>.toFullscreen(index: Int = value.currentImageIndex) = modifyState {
+    copy(
+        currentImageIndex = index,
+        screen = ScreenState.FullScreen
     )
-
-    override fun notifyRefreshUnavailable() = showPopUpMessage(
-        """
-            ${localization.noInternet}
-            ${localization.refreshUnavailable}
-        """.trimIndent()
-    )
-}
-
-interface Localization {
-    val back: String
-    val appName: String
-    val loading: String
-    val repoInvalid: String
-    val repoEmpty: String
-    val noInternet: String
-    val loadImageUnavailable: String
-    val lastImage: String
-    val firstImage: String
-    val picture: String
-    val size: String
-    val pixels: String
-    val refreshUnavailable: String
 }
