+++ experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/view/MainScreen.kt
@@ -7,16 +7,13 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import example.imageviewer.Dependencies
 import example.imageviewer.model.*
+import example.imageviewer.model.State
 import example.imageviewer.style.*
 import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.orEmpty
@@ -35,13 +32,10 @@ internal fun MainScreen(state: MutableState<State>, dependencies: Dependencies) 
                     picture = picture,
                     image = state.value.miniatures[picture],
                     onClickSelect = {
-                        state.setMainImage(picture, dependencies)
+                        state.setSelectedIndex(i)
                     },
                     onClickFullScreen = {
-                        state.value = state.value.copy(
-                            currentImageIndex = i,
-                            screen = ScreenState.FullScreen
-                        )
+                        state.toFullscreen(i)
                     },
                     onClickInfo = {
                         dependencies.notification.notifyImageData(picture)
@@ -59,10 +53,10 @@ internal fun MainScreen(state: MutableState<State>, dependencies: Dependencies) 
 @Composable
 private fun TopContent(state: MutableState<State>, dependencies: Dependencies) {
     TitleBar(state, dependencies)
-//    if (LocalConfiguration.current.orientation == Configuration.ORIENTATION_PORTRAIT) {//todo
-    PreviewImage(state)
-    Spacer(modifier = Modifier.height(10.dp))
-//    }
+    if (needShowPreview()) {
+        PreviewImage(state = state, getImage = { dependencies.imageRepository.loadContent(it.bigUrl) })
+        Spacer(modifier = Modifier.height(5.dp))
+    }
     Spacer(modifier = Modifier.height(5.dp))
 }
 
@@ -83,115 +77,14 @@ private fun TitleBar(state: MutableState<State>, dependencies: Dependencies) {
                     modifier = Modifier.padding(end = 20.dp).align(Alignment.CenterVertically),
                     shape = CircleShape
                 ) {
-                    Clickable(
-                        onClick = {
-                            if (state.value.isContentReady) {
-                                state.refresh(dependencies)
-                            }
+                    Image(
+                        bitmap = resource("refresh.png").rememberImageBitmap().orEmpty(),
+                        contentDescription = null,
+                        modifier = Modifier.size(35.dp).clickable {
+                            state.refresh(dependencies)
                         }
-                    ) {
-                        Image(
-                            bitmap = resource("refresh.png").rememberImageBitmap().orEmpty(),
-                            contentDescription = null,
-                            modifier = Modifier.size(35.dp)
-                        )
-                    }
+                    )
                 }
             }
         })
 }
-
-@OptIn(ExperimentalResourceApi::class)
-@Composable
-internal fun PreviewImage(state: MutableState<State>) {
-    Clickable(onClick = {
-        state.value = state.value.copy(
-            screen = ScreenState.FullScreen
-        )
-    }) {
-        Card(
-            backgroundColor = DarkGray,
-            modifier = Modifier.height(250.dp),
-            shape = RectangleShape,
-            elevation = 1.dp
-        ) {
-            Image(
-                bitmap = state.value.mainImage ?: resource("empty.png").rememberImageBitmap().orEmpty(),
-                contentDescription = null,
-                modifier = Modifier
-                    .fillMaxWidth().padding(start = 1.dp, top = 1.dp, end = 1.dp, bottom = 5.dp),
-                contentScale = ContentScale.Fit
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalResourceApi::class)
-@Composable
-internal fun Miniature(
-    picture: Picture,
-    image: ImageBitmap?,
-    onClickSelect: () -> Unit,
-    onClickFullScreen: () -> Unit,
-    onClickInfo: () -> Unit,
-) {
-    Card(
-        backgroundColor = MiniatureColor,
-        modifier = Modifier.padding(start = 10.dp, end = 10.dp).height(70.dp)
-            .fillMaxWidth()
-            .clickable {
-                onClickSelect()
-            },
-        shape = RectangleShape,
-        elevation = 2.dp
-    ) {
-        Row(modifier = Modifier.padding(end = 30.dp)) {
-            Clickable(
-                onClick = {
-                    onClickFullScreen()
-                }
-            ) {
-                val modifier = Modifier.height(70.dp)
-                    .width(90.dp)
-                    .padding(start = 1.dp, top = 1.dp, end = 1.dp, bottom = 1.dp)
-                if (image != null) {
-                    Image(
-                        image,
-                        contentDescription = null,
-                        modifier = modifier,
-                        contentScale = ContentScale.Crop
-                    )
-                } else {
-                    CircularProgressIndicator(
-                        modifier = modifier,
-                        color = DarkGreen
-                    )
-                }
-            }
-            Text(
-                text = picture.name,
-                color = Foreground,
-                modifier = Modifier.weight(1f).align(Alignment.CenterVertically).padding(start = 16.dp),
-                style = MaterialTheme.typography.body1
-            )
-
-            Clickable(
-                modifier = Modifier.height(70.dp)
-                    .width(30.dp),
-                onClick = {
-                    onClickInfo()
-                }
-            ) {
-                Image(
-                    resource("dots.png").rememberImageBitmap().orEmpty(),
-                    contentDescription = null,
-                    modifier = Modifier.height(70.dp)
-                        .width(30.dp)
-                        .padding(start = 1.dp, top = 25.dp, end = 1.dp, bottom = 25.dp),
-                    contentScale = ContentScale.FillHeight
-                )
-            }
-        }
-    }
-}
-
