+++ experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/view/ScalableImage.common.kt
@@ -5,18 +5,21 @@ import androidx.compose.foundation.gestures.detectTransformGestures
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
-
+import example.imageviewer.model.ScalableState
+import example.imageviewer.model.addDragAmount
+import example.imageviewer.model.addScale
+import example.imageviewer.model.setScale
 
 expect fun Modifier.addUserInput(state: MutableState<ScalableState>): Modifier
-
 
 fun Modifier.addTouchUserInput(state: MutableState<ScalableState>): Modifier =
     pointerInput(Unit) {
         detectTransformGestures { _, pan, zoom, _ ->
-            state.value = state.value.addDragAmount(pan).addScale(zoom - 1f)
+            state.addDragAmount(pan)
+            state.addScale(zoom - 1f)
         }
     }.pointerInput(Unit) {
-        detectTapGestures(onDoubleTap = {
-            state.value = state.value.copy(scale = 1f)
-        })
+        detectTapGestures(
+            onDoubleTap = { state.setScale(1f) }
+        )
     }
