+++ experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/utils/GraphicsMath.kt
@@ -1,22 +1,12 @@
 package example.imageviewer.utils
 
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import java.awt.Dimension
-import java.awt.Graphics2D
-import java.awt.Rectangle
 import java.awt.Toolkit
 import java.awt.image.BufferedImage
-import java.io.ByteArrayOutputStream
-import java.io.File
-import java.io.IOException
-import javax.imageio.ImageIO
-import java.awt.image.BufferedImageOp
 import java.awt.image.ConvolveOp
 import java.awt.image.Kernel
-import kotlin.math.pow
-import kotlin.math.roundToInt
 
 fun scaleBitmapAspectRatio(
     bitmap: BufferedImage,
@@ -39,22 +29,6 @@ fun scaleBitmapAspectRatio(
     graphics.dispose()
 
     return result
-}
-
-fun getDisplayBounds(bitmap: BufferedImage, windowSize: DpSize): Rectangle {
-
-    val boundW: Float = windowSize.width.value
-    val boundH: Float = windowSize.height.value
-
-    val ratioX: Float = bitmap.width / boundW
-    val ratioY: Float = bitmap.height / boundH
-
-    val ratio: Float = if (ratioX > ratioY) ratioX else ratioY
-
-    val resultW = (boundW * ratio)
-    val resultH = (boundH * ratio)
-
-    return Rectangle(0, 0, resultW.toInt(), resultH.toInt())
 }
 
 fun applyGrayScaleFilter(bitmap: BufferedImage): BufferedImage {
@@ -109,85 +83,6 @@ fun applyBlurFilter(bitmap: BufferedImage): BufferedImage {
         result.width - radius * 2,
         result.height - radius * 2
     )
-}
-
-fun cropImage(bitmap: BufferedImage, crop: Rectangle) : BufferedImage {
-    return bitmap.getSubimage(crop.x, crop.y, crop.width, crop.height)
-}
-
-fun cropBitmapByScale(
-    bitmap: BufferedImage,
-    size: DpSize,
-    scale: Float,
-    offset: Offset,
-): BufferedImage {
-    val crop = cropBitmapByBounds(
-        bitmap,
-        getDisplayBounds(bitmap, size),
-        size,
-        scale,
-        offset,
-    )
-    return cropImage(
-        bitmap,
-        Rectangle(crop.x, crop.y, crop.width - crop.x, crop.height - crop.y)
-    )
-}
-
-fun cropBitmapByBounds(
-    bitmap: BufferedImage,
-    bounds: Rectangle,
-    size: DpSize,
-    scaleFactor: Float,
-    offset:Offset
-): Rectangle {
-
-    if (scaleFactor <= 1f) {
-        return Rectangle(0, 0, bitmap.width, bitmap.height)
-    }
-
-    var scale = scaleFactor.toDouble().pow(1.4)
-
-    var boundW = (bounds.width / scale).roundToInt()
-    var boundH = (bounds.height / scale).roundToInt()
-
-    scale *= size.width.value / bounds.width.toDouble()
-
-    val offsetX = offset.x / scale
-    val offsetY = offset.y / scale
-
-    if (boundW > bitmap.width) {
-        boundW = bitmap.width
-    }
-    if (boundH > bitmap.height) {
-        boundH = bitmap.height
-    }
-
-    val invisibleW = bitmap.width - boundW
-    var leftOffset = (invisibleW / 2.0 - offsetX).roundToInt()
-
-    if (leftOffset > invisibleW) {
-        leftOffset = invisibleW
-//        drag.getAmount().x = -((invisibleW / 2.0) * scale).roundToInt().toFloat()
-    }
-    if (leftOffset < 0) {
-//        drag.getAmount().x = ((invisibleW / 2.0) * scale).roundToInt().toFloat()
-        leftOffset = 0
-    }
-
-    val invisibleH = bitmap.height - boundH
-    var topOffset = (invisibleH / 2 - offsetY).roundToInt()
-
-    if (topOffset > invisibleH) {
-        topOffset = invisibleH
-//        drag.getAmount().y = -((invisibleH / 2.0) * scale).roundToInt().toFloat()
-    }
-    if (topOffset < 0) {
-//        drag.getAmount().y = ((invisibleH / 2.0) * scale).roundToInt().toFloat()
-        topOffset = 0
-    }
-
-    return Rectangle(leftOffset, topOffset, leftOffset + boundW, topOffset + boundH)
 }
 
 fun getPreferredWindowSize(desiredWidth: Int, desiredHeight: Int): DpSize {
