+++ experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/view/PreviewImage.desktop.kt
@@ -1,0 +1,6 @@
+package example.imageviewer.view
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal actual fun needShowPreview(): Boolean = true
