+++ experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/view/ScalaImage.desktop.kt
@@ -2,18 +2,18 @@ package example.imageviewer.view
 
 import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.runtime.MutableState
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.input.key.*
 import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.pointerInput
+import example.imageviewer.model.ScalableState
+import example.imageviewer.model.addDragAmount
+import example.imageviewer.model.addScale
 
-@OptIn(ExperimentalComposeUiApi::class)
-actual fun Modifier.addUserInput(state:MutableState<ScalableState>):Modifier =
+actual fun Modifier.addUserInput(state: MutableState<ScalableState>): Modifier =
     pointerInput(Unit) {
         detectDragGestures { change, dragAmount: Offset ->
-            state.value = state.value.addDragAmount(dragAmount)
+            state.addDragAmount(dragAmount)
             change.consume()
         }
     }.pointerInput(Unit) {
@@ -22,25 +22,8 @@ actual fun Modifier.addUserInput(state:MutableState<ScalableState>):Modifier =
                 val event = awaitPointerEvent()
                 if (event.type == PointerEventType.Scroll) {
                     val delta = event.changes.getOrNull(0)?.scrollDelta ?: Offset.Zero
-                    state.value = state.value.addScale(delta.y / 100)
+                    state.addScale(delta.y / 100)
                 }
             }
         }
-    }.onPreviewKeyEvent {
-        if (it.type == KeyEventType.KeyUp) {
-            when (it.key) {
-                Key.I, Key.Plus, Key.Equals -> {
-                    state.value = state.value.addScale(0.2f)
-                }
-
-                Key.O, Key.Minus -> {
-                    state.value = state.value.addScale(-0.2f)
-                }
-
-                Key.R -> {
-                    state.value = state.value.copy(scale = 1f)
-                }
-            }
-        }
-        false
     }
