+++ experimental/examples/imageviewer/shared/src/iosMain/kotlin/example/imageviewer/main.ios.kt
@@ -1,10 +1,5 @@
 package example.imageviewer
 
-/*
- * Copyright 2020-2022 JetBrains s.r.o. and respective authors and developers.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
- */
-
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.height
