+++ experimental/examples/imageviewer/shared/src/iosMain/kotlin/example/imageviewer/view/Tooltip.ios.kt
@@ -3,9 +3,10 @@ package example.imageviewer.view
 import androidx.compose.runtime.Composable
 
 @Composable
-internal actual fun Tooltip( //todo ios
+internal actual fun Tooltip(
     text: String,
     content: @Composable () -> Unit
 ) {
+    //No tooltip for iOS
     content()
 }
