+++ experimental/examples/imageviewer/shared/src/iosMain/kotlin/example/imageviewer/view/addUserInput.kt
@@ -2,6 +2,7 @@ package example.imageviewer.view
 
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
+import example.imageviewer.model.ScalableState
 
 actual fun Modifier.addUserInput(state: MutableState<ScalableState>): Modifier =
     addTouchUserInput(state)
